Fusion des deux composants de la navbar en un seul composant et suppression des fichiers en conséquence
