+++ frontend/src/pages/BackOffice.jsx
@@ -1,19 +1,12 @@
 import React from "react";
-import NavBarBackOfficeDesktop from "@components/NavBarBackOfficeDesktop";
-import NavBarBackOfficeMobile from "@components/NavBarBackOfficeMobile";
+import NavBarBackOffice from "@components/NavBarBackOffice";
 import "@style/App.css";
 import "@style/BackOffice.css";
 
 function BackOffice() {
   return (
     <div className="arch-backoffice">
-      <div className="navbar-desktop-backoffice">
-        <NavBarBackOfficeDesktop />
-      </div>
-
-      <div className="navbar-mobile-backoffice">
-        <NavBarBackOfficeMobile />
-      </div>
+      <NavBarBackOffice />
       <div className="app-main-container">
         {/* LES ROUTES SERONT ICI */} ICI S'AFFICHERONT LES PAGES
       </div>
